# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationResponse": {
        "additionalProperties": false,
        "description": "One source, and whether it is currently reading.\n\nDisconnected integrations are returned rather than filtered out: a gap in the\nfeed is explained by \"GitHub was disconnected on the 4th\" and unexplained by\nsilence.",
        "properties": {
          "account": {
            "title": "Account",
            "type": "string"
          },
          "authorisedBy": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorisedby"
          },
          "connectedAt": {
            "format": "date-time",
            "title": "Connectedat",
            "type": "string"
          },
          "disconnectedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disconnectedat"
          },
          "health": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health"
          },
          "installationId": {
            "title": "Installationid",
            "type": "integer"
          },
          "lastSuccessfulSyncAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastsuccessfulsyncat"
          },
          "revokedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revokedat"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "suspended": {
            "default": false,
            "title": "Suspended",
            "type": "boolean"
          }
        },
        "required": ["source", "account", "installationId", "connectedAt"],
        "title": "IntegrationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "CAIRN's HTTP API.\n\nErrors are RFC 9457 problem documents (`application/problem+json`) with a stable\n`type` URI to branch on. Authentication is a `HttpOnly` session cookie issued by\n`POST /v1/auth/login`; browser clients must send credentials and never need to\nread the cookie.",
    "title": "CAIRN API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{workspace_id}/integrations": {
      "get": {
        "description": "Every integration, connected or not, and its state.\n\nReadable by every member rather than by administrators only. What is\nconnected decides what CAIRN can see about the person reading \u2014 the same\nreasoning that puts the source list on the notification screen. Hiding it\nbehind a role would mean a Viewer had to ask permission to find out what was\nbeing read about them.\n\nDisconnected installations are listed, not omitted: a gap in the feed is\nexplained by \"GitHub was disconnected on the 4th\" and unexplained by silence.\n\nRead through the tenant-scoped connection even though the write below is\nplatform-side. The application role has `SELECT` on this table behind a\nrow-level-security policy, so the isolation is enforced by the database\nrather than by the `WHERE` clause being remembered.",
        "operationId": "list_integrations_v1_workspaces__workspace_id__integrations_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cairn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cairn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationResponse"
                  },
                  "title": "Response List Integrations V1 Workspaces  Workspace Id  Integrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such workspace, or you are not a member."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What this workspace has connected",
        "tags": ["administration"]
      }
    },
...
    "/v1/workspaces/{workspace_id}/integrations/github/{installation_id}": {
      "delete": {
        "description": "Stop capture from one installation.\n\n**Marked, not deleted**, and the two differ in what they promise. Marking\nstops every future delivery being processed \u2014 the ingestion path checks this\nalready, because a suspended installation that kept delivering was treated as\na consent problem rather than a bug. Deleting the row would additionally\nerase the record that the integration ever existed, which turns the months of\nactivity it produced into facts with no explanation of where they came from.\n\n**What was already captured stays**, and the interface says so rather than\nletting an administrator discover it. Disconnecting is \"stop reading\", not\n\"forget what you read\": the second is a deletion request, it applies to\neveryone's shared history, and it is not a side effect of a button labelled\n*Disconnect*.\n\nThis does not uninstall the GitHub App. CAIRN cannot revoke somebody else's\ninstallation, and pretending otherwise would leave an administrator believing\nthey had removed an access grant that is still live on GitHub's side.",
        "operationId": "disconnect_github_v1_workspaces__workspace_id__integrations_github__installation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "installation_id",
            "required": true,
            "schema": {
              "title": "Installation Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cairn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cairn Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Requires permission to disconnect integrations."
          },
          "404": {
            "description": "No such installation in this workspace."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop capturing activity from a GitHub installation",
        "tags": ["administration"]
      }
    },
...
    "/v1/workspaces/{workspace_id}/notifications": {
      "get": {
        "description": "Notification per person; opt-outs only as a number.\n\n**This is the asymmetry, and it is the most considered decision in the\nmodule.** md/15 \u00a74.2 describes one screen showing \"who has been notified, who\nhas opted out\". The first half is named per person and the second is not, and\nthe two halves are different kinds of fact:\n\n- **Notification is the employer's obligation**, owed to each person\n  individually before capture begins, with no regional exception. An Owner\n  who cannot see that Priya has not been notified cannot discharge it, and\n  cannot evidence it when a works council asks. So it is named.\n- **An opt-out is the person's own decision about their own record.** A list\n  of names beside \"opted out\" is a list of employees who declined to be\n  recorded, handed to the person who writes their review. It does not matter\n  that no reasonable manager would misuse it; what matters is that a person\n  deciding whether to opt out would have to weigh that possibility, which\n  turns a privacy control into a career calculation and produces a low\n  opt-out rate that means nothing.\n\nSo the rate is reported and the names are not. That is also the number\nmd/11 \u00a77 makes the product's trust barometer and md/13 makes a phase gate \u2014\nand a rate is exactly what a gate needs, where a list is not.\n\nThe permission is `WORKSPACE_SETTINGS` rather than `CONTENT_READ`: whether a\ncolleague has been notified is compliance administration rather than\nsomething everyone needs, and this endpoint names people.",
        "operationId": "notification_status_v1_workspaces__workspace_id__notifications_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cairn_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cairn Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Requires permission to change workspace settings."
          },
          "404": {
            "description": "No such workspace, or you are not a member."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Who has been told their activity may be captured",
        "tags": ["administration"]
      }
    },
...
  }
}